Section1-WeatherHolidaysPipeline/pyproject.toml
[project]
name = "weather_holidays_pipeline"
version = "0.1.0"
requires-python = ">=3.10"

dependencies = [
    "requests>=2.32.5",
    "pandas>=2.3.2",
    "python-dotenv>=1.1.1"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.1",
    "pytest-mock>=3.14.1"
]